+++ src/client.ts
@@ -407,7 +407,7 @@ class Client {
         'x-sdk-platform': 'console',
         'x-sdk-language': 'web',
         'x-sdk-version': '9.0.0',
-        'X-Appwrite-Response-Format': '1.9.0',
+        'X-Appwrite-Response-Format': '1.9.1',
     };
 
     /**