+++ src/enums/platform-linux-type.ts
@@ -1,7 +1,0 @@
-export enum PlatformLinuxType {
-    Windows = 'windows',
-    Apple = 'apple',
-    Android = 'android',
-    Linux = 'linux',
-    Web = 'web',
-}