+++ src/enums/platform-windows-type.ts
@@ -1,7 +1,0 @@
-export enum PlatformWindowsType {
-    Windows = 'windows',
-    Apple = 'apple',
-    Android = 'android',
-    Linux = 'linux',
-    Web = 'web',
-}